requirements.txt
gunicorn==20.0.4
Flask==2.0.1
line-bot-sdk==1.16.0
itsdangerous==2.0.1
Flask_Cors==3.0.2
